Move `list_py_file_paths` test to the right file (#45521)
(cherry picked from commit c29b6e8e0ba739aa66ddf371cd3fa0c79e3ce651)
